+++ packages/@coorpacademy-components/src/atom/review-presentation/prop-types.ts
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types';
-
+import {GestureResponderEvent} from 'react-native';
 const levelItem = PropTypes.shape({
   text: PropTypes.string,
   tooltipText: PropTypes.string
@@ -18,3 +18,19 @@ const propTypes = {
 };
 
 export default propTypes;
+
+export const OnboardingPropsTypes = {
+  onPress: PropTypes.func
+};
+
+export const TipPropsTypes = {
+  text: PropTypes.string,
+  Icon: PropTypes.any
+};
+
+export type OnboardingProps = {onPress: (event: GestureResponderEvent) => any};
+
+export type TipProps = {
+  text: string | undefined;
+  Icon: any;
+};
