+++ packages/@coorpacademy-components/src/template/app-review/onboarding/index.native.tsx
@@ -11,7 +11,7 @@ import {useTemplateContext} from '../template-context';
 import {Theme} from '../../../variables/theme.native';
 import {HEADER_HEIGHT} from '../../../organism/header-v2/index.native';
 import Touchable from '../../../hoc/touchable/index.native';
-import {OnboardingProps, TipProps} from './prop-types';
+import {OnboardingProps, TipProps} from '../../../atom/review-presentation/prop-types';
 
 type StyleSheetType = {
   container: ViewStyle;
